# tools/bin/ut: accelerate building test binary