--- a/tools/check/ut.go
+++ b/tools/check/ut.go
@@ -133,12 +133,10 @@ func cmdBuild(args ...string) bool {
 
 	// build all packages
 	if len(args) == 0 {
-		for _, pkg := range pkgs {
-			err := buildTestBinary(pkg)
-			if err != nil {
-				fmt.Println("build package error", pkg, err)
-				return false
-			}
+		err := buildTestBinaryMulti(pkgs)
+		if err != nil {
+			fmt.Println("build package error", pkgs, err)
+			return false
 		}
 		return true
 	}
@@ -163,16 +161,16 @@ func cmdRun(args ...string) bool {
 		return false
 	}
 	tasks := make([]task, 0, 5000)
+	start := time.Now()
 	// run all tests
 	if len(args) == 0 {
-		for _, pkg := range pkgs {
-			fmt.Println("handling package", pkg)
-			err := buildTestBinary(pkg)
-			if err != nil {
-				fmt.Println("build package error", pkg, err)
-				return false
-			}
+		err := buildTestBinaryMulti(pkgs)
+		if err != nil {
+			fmt.Println("build package error", pkgs, err)
+			return false
+		}
 
+		for _, pkg := range pkgs {
 			exist, err := testBinaryExist(pkg)
 			if err != nil {
 				fmt.Println("check test binary existance error", err)
@@ -248,7 +246,7 @@ func cmdRun(args ...string) bool {
 		}
 		tasks = tmp
 	}
-	fmt.Println("building task finish...", len(tasks))
+	fmt.Printf("building task finish, count=%d, takes=%v\n", len(tasks), time.Since(start))
 
 	numactl := numactlExist()
 	taskCh := make(chan task, 100)
@@ -446,7 +444,27 @@ func buildTestBinary(pkg string) error {
 		return withTrace(err)
 	}
 	return nil
+}
 
+// buildTestBinaryMulti is much faster than build the test packages one by one.
+func buildTestBinaryMulti(pkgs []string) error {
+       // go test --exec=xprog -cover -vet=off --count=0 $(pkgs)
+       xprogPath := path.Join(workDir, "tools/bin/xprog")
+       packages := make([]string, 0, len(pkgs))
+       for _, pkg := range pkgs {
+               packages = append(packages, path.Join(modulePath, pkg))
+       }
+
+       var cmd *exec.Cmd
+	cmd = exec.Command("go", "test", "--exec", xprogPath, "-vet", "off", "-count", "0")
+       cmd.Args = append(cmd.Args, packages...)
+       cmd.Dir = workDir
+       cmd.Stdout = os.Stdout
+       cmd.Stderr = os.Stderr
+       if err := cmd.Run(); err != nil {
+               return withTrace(err)
+       }
+       return nil
 }
 
 func testBinaryExist(pkg string) (bool, error) {