<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #31880
A subtask for #30822

Problem Summary:

Build the test binary one by one is painfully slow.

Before change:
```
6m39.360986586s
```

After change:
```
1m56.399393108s
```

### What is changed and how it works?

`go test --exec=xprog`, the Go compile toolchain would call your `xprog` program passing the test binary path as the argument...

So you can do this trick, write you own `xprog`, and move the test binary to some place else, save it for later use.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)

Now you can use `go test --exec=xprog` to build all test binaries at once.

```
genius@genius-System-Product-Name:~/project/src/github.com/pingcap/tidb$ pwd
/home/genius/project/src/github.com/pingcap/tidb


genius@genius-System-Product-Name:~/project/src/github.com/pingcap/tidb$ make tools/bin/xprog
cd tools/check; \
GO111MODULE=on go build -o ../bin/xprog xprog.go

genius@genius-System-Product-Name:~/project/src/github.com/pingcap/tidb$ go test --exec=/home/genius/project/src/github.com/pingcap/tidb/tools/bin/xprog ./...
ok      github.com/pingcap/tidb/bindinfo        0.062s
ok      github.com/pingcap/tidb/br/cmd/br       0.048s
ok      github.com/pingcap/tidb/br/cmd/tidb-lightning   0.027s
ok      github.com/pingcap/tidb/br/cmd/tidb-lightning-ctl       0.007s
ok      github.com/pingcap/tidb/br/pkg/backup   0.010s
ok      github.com/pingcap/tidb/br/pkg/checksum 0.003s
ok      github.com/pingcap/tidb/br/pkg/conn     0.020s
?       github.com/pingcap/tidb/br/pkg/errors   [no test files]
?       github.com/pingcap/tidb/br/pkg/glue     [no test files]
?       github.com/pingcap/tidb/br/pkg/gluetidb [no test files]
ok      github.com/pingcap/tidb/br/pkg/gluetikv 0.028s
?       github.com/pingcap/tidb/br/pkg/httputil [no test files]
ok      github.com/pingcap/tidb/br/pkg/kv       0.017s
ok      github.com/pingcap/tidb/br/pkg/lightning        0.028s
ok      github.com/pingcap/tidb/br/pkg/lightning/backend        0.060s
ok      github.com/pingcap/tidb/br/pkg/lightning/backend/importer       0.020s
ok      github.com/pingcap/tidb/br/pkg/lightning/backend/kv     0.064s
ok      github.com/pingcap/tidb/br/pkg/lightning/backend/local  0.014s
?       github.com/pingcap/tidb/br/pkg/lightning/backend/noop   [no test files]
ok      github.com/pingcap/tidb/br/pkg/lightning/backend/tidb   0.049s
ok      github.com/pingcap/tidb/br/pkg/lightning/checkpoints    0.067s
?       github.com/pingcap/tidb/br/pkg/lightning/checkpoints/checkpointspb      [no test files]
ok      github.com/pingcap/tidb/br/pkg/lightning/common 0.021s
ok      github.com/pingcap/tidb/br/pkg/lightning/config 0.002s
ok      github.com/pingcap/tidb/br/pkg/lightning/errormanager   0.017s
?       github.com/pingcap/tidb/br/pkg/lightning/glue   [no test files]
ok      github.com/pingcap/tidb/br/pkg/lightning/log    0.002s
?       github.com/pingcap/tidb/br/pkg/lightning/manual [no test files]
ok      github.com/pingcap/tidb/br/pkg/lightning/metric 0.013s
ok      github.com/pingcap/tidb/br/pkg/lightning/mydump 0.034s
ok      github.com/pingcap/tidb/br/pkg/lightning/restore        0.015s
ok      github.com/pingcap/tidb/br/pkg/lightning/tikv   0.009s
ok      github.com/pingcap/tidb/br/pkg/lightning/verification   0.030s
?       github.com/pingcap/tidb/br/pkg/lightning/web    [no test files]
ok      github.com/pingcap/tidb/br/pkg/lightning/worker 0.017s
...

genius@genius-System-Product-Name:~/project/src/github.com/pingcap/tidb$ find . -name '*.test.bin'
./tablecodec/rowindexcodec/rowindexcodec.test.bin
./tablecodec/tablecodec.test.bin
./config/config.test.bin
./meta/autoid/autoid.test.bin
./meta/meta.test.bin
./metrics/metrics.test.bin
./privilege/privileges/privileges.test.bin
./executor/executor.test.bin
./executor/seqtest/seqtest.test.bin
./executor/aggfuncs/aggfuncs.test.bin
./executor/oomtest/oomtest.test.bin
./owner/owner.test.bin
./kv/kv.test.bin
./br/pkg/task/task.test.bin
./br/pkg/checksum/checksum.test.bin
./br/pkg/version/build/build.test.bin
./br/pkg/version/version.test.bin
./br/pkg/mock/mock.test.bin
./br/pkg/gluetikv/gluetikv.test.bin
```

- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```